@@ -1,4 +1,4 @@
-from typing import ClassVar, List
+from typing import ClassVar
 
 from wagtail.api.v2.views import BaseAPIViewSet
 
@@ -7,7 +7,7 @@ from .models import Author
 
 class AuthorsAPIViewSet(BaseAPIViewSet):
     model = Author
-    listing_default_fields: ClassVar[List[str]] = [
+    listing_default_fields: ClassVar[list[str]] = [
         'id',
         'detail_url',
         'name',
