@@ -1,4 +1,4 @@
-from typing import ClassVar, List
+from typing import ClassVar
 
 from wagtail.api.v2.views import BaseAPIViewSet
 
@@ -8,7 +8,7 @@ from .models import ExhibitPage
 class ExhibitsAPIViewSet(BaseAPIViewSet):
     model = ExhibitPage
 
-    listing_default_fields: ClassVar[List[str]] = [
+    listing_default_fields: ClassVar[list[str]] = [
         *BaseAPIViewSet.listing_default_fields,
         'title',
         'body',
