@@ -1,4 +1,4 @@
-from typing import ClassVar, List
+from typing import ClassVar
 
 from wagtail.admin.panels import FieldPanel
 from wagtail.api import APIField
@@ -9,9 +9,9 @@ from wagtail.models import Page
 class HomePage(Page):
     body = RichTextField(blank=True)
 
-    content_panels: ClassVar[List[FieldPanel]] = [
+    content_panels: ClassVar[list[FieldPanel]] = [
         *Page.content_panels,
         FieldPanel('body', classname='full'),
     ]
 
-    api_fields: ClassVar[List[APIField]] = [APIField('body')]
+    api_fields: ClassVar[list[APIField]] = [APIField('body')]
